requirements.txt
absl-py==2.1.0
alembic==1.15.2
annotated-types==0.6.0
ansiwrap==0.8.4
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.6
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Automat==20.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.0.0
blinker==1.4
blis==0.7.11
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
build==0.10.0
catalogue==2.0.10
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
cloud-init==24.4.1
cloudpathlib==0.16.0
cmake==3.27.2
colorama==0.4.4
colorlog==6.9.0
comm==0.1.4
command-not-found==0.3
confection==0.1.4
ConfigArgParse==1.7
configobj==5.0.6
connection-pool==0.0.3
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.8
datrie==0.8.2
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
dpath==2.1.6
duplicity==0.8.21
et-xmlfile==1.0.1
exceptiongroup==1.1.3
executing==1.2.0
fail2ban==0.11.2
fastai==2.7.15
fastcore==1.5.33
fastdownload==0.0.7
fasteners==0.14.1
fastjsonschema==2.18.0
fastprogress==1.0.3
filelock==3.12.2
flatbuffers==24.3.25
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
fsspec==2023.9.2
future==0.18.2
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
google-pasta==0.2.0
gpg==1.16.0
greenlet==3.1.1
grpcio==1.62.1
h5py==3.11.0
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
imageio==2.31.1
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
isort==5.12.0
jdcal==1.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
json5==0.9.14
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==3.3.3
keyring==23.5.0
kiwisolver==1.3.2
langcodes==3.4.0
language-selector==0.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lazy_loader==0.3
ldap3==2.9.1
libclang==18.1.1
lit==16.0.6
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mercurial==6.1.1
mistune==3.0.1
ml-dtypes==0.3.2
monai==1.3.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
murmurhash==1.0.10
mypy==1.5.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.1
nose==1.3.7
notebook==7.0.2
notebook_shim==0.2.3
numba==0.55.1
numexpr==2.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
opencv-python==4.11.0.86
openpyxl==3.0.9
opt-einsum==3.3.0
optree==0.11.0
optuna==4.2.1
overrides==7.4.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plac==1.3.5
platformdirs==3.10.0
pluggy==0.13.0
ply==3.11
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
py==1.10.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.17.5
pymacaroons==0.13.0
pymemcache==4.0.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-json-logger==2.0.7
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
reportlab==3.6.8
requests==2.31.0
reretry==0.11.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.9.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==2.25.1
service-identity==18.1.0
setproctitle==1.3.5
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
snakemake==7.32.3
sniffio==1.3.0
sos==4.8.2
soupsieve==2.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.2
stopit==1.1.2
sympy==1.9
systemd-python==234
tables==3.7.0
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.17.1
texttable==1.6.4
textwrap3==0.9.2
thinc==8.2.3
threadpoolctl==3.2.0
throttler==1.2.2
tifffile==2023.8.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.12.1
toposort==1.10
torch==2.4.1
torchvision==0.19.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.9.0
triton==3.0.0
Twisted==22.1.0
typer==0.9.4
typing_extensions==4.13.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ultralytics==8.3.104
ultralytics-thop==2.0.14
unattended-upgrades==0.1
unet==0.7.7
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==2.3.0
usb-creator==0.3.7
virtualenv==20.24.3
wadllib==1.3.6
wandb==0.19.9
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wrapt==1.15.0
xdg==5
xkit==0.0.0
xlwt==1.3.0
yte==1.5.1
zipp==1.0.0
zope.interface==5.4.0
